Numbers still set to NULL when set to empty string
Ran into an issue recently where a nullable integer field, never set on creation of a new model, was causing problems after a JSON submit back to the server. Turns out that serialization methods on the client was turning all null things into empty strings.

This change allows for that by making the Number comparison a little more strict. We now generate NULL for the query even if the value coerces to zero (which empty string does).