Force Postgres to acknowledge our UTC dates
Got this error in the tests:

  datatypes should keep types when get read data from db:
  actual:    Fri Jul 19 2013 18:04:41 GMT-0700 (PDT)
  expected:  Fri Jul 19 2013 11:04:41 GMT-0700 (PDT)

We're not roundtripping from the database properly. By attempting to save into the database in UTC, but not telling the server the value has a time zone, we save in the server's local time. From the postgres docs: "PostgreSQL never examines the content of a literal string before determining its type, and therefore will treat both of the above as timestamp without time zone." (http://www.postgresql.org/docs/9.2/static/datatype-datetime.html)

So, this change forces the database to acknowledge a timezone on the provided string. That strange .substring(1) construction removes the 'E' prefix added by escape(). We escape the result of toISOString() in case we get an object that returns malicious things from it's toISOString() method.


Regarding alternatives: before the recent change to save UTC dates, we were assuming that the local app is running in the same time zone as the postgres server. That has worked for me in the past, but this solution seems to be the better option. Of course, I don't like adding that long string to every single saved timestamp, but that's how it is in the Postgres world.